[Refactor] engine 패키지 구조 변경 및 usecase별 port, adaptor 추가 (#296)
* refactor: url mapping 수정

@RequestMapping 제거

* feat: GetTrimsPort, GetTrimsPort 추가

* refactor: Port 변경

EnginePort에서 GetEginesPort로 변경

* chore: 응답 생성 메서드 위치 수정

* chore: 요청 DTO 네이밍 수정
