+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/engine/usecase/get_engines/GetEnginesUseCase.java
@@ -1,10 +1,9 @@
-package softeer.be_my_car_master.api.engine.usecase;
+package softeer.be_my_car_master.api.engine.usecase.get_engines;
 
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.engine.dto.response.GetEnginesResponse;
-import softeer.be_my_car_master.api.engine.usecase.port.EnginePort;
 import softeer.be_my_car_master.domain.engine.Engine;
 import softeer.be_my_car_master.global.annotation.UseCase;
 
@@ -12,10 +11,10 @@ import softeer.be_my_car_master.global.annotation.UseCase;
 @RequiredArgsConstructor
 public class GetEnginesUseCase {
 
-	private final EnginePort enginePort;
+	private final GetEnginesPort getEnginesPort;
 
 	public GetEnginesResponse execute(Long trimId) {
-		List<Engine> selectableEngines = enginePort.findSelectableEnginesByTrimId(trimId);
-		return GetEnginesResponse.from(selectableEngines);
+		List<Engine> engines = getEnginesPort.findEnginesByTrim(trimId);
+		return GetEnginesResponse.from(engines);
 	}
 }
